% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{benchmark_causal}
\alias{benchmark_causal}
\title{Causal Benchmark Results}
\format{
A list containing:
\describe{
\item{rLifting_Time_Avg}{Average time (seconds) for rLifting.}
\item{Wavethresh_Naive_Time}{Time (seconds) for naive sliding window.}
\item{Speedup_Factor}{Ratio of Naive Time to rLifting Time.}
}
}
\usage{
data(benchmark_causal)
}
\description{
Comparison of execution time between \code{rLifting}'s optimized causal mode and
a naive sliding-window implementation using \code{wavethresh}.
}
\keyword{datasets}
