% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{benchmark_offline}
\alias{benchmark_offline}
\title{Offline Benchmark Results}
\format{
A data frame with the following columns:
\describe{
\item{Pkg}{Package name.}
\item{Time}{Execution time in seconds.}
\item{MSE}{Mean Squared Error.}
}
}
\usage{
data(benchmark_offline)
}
\description{
Comparison of execution time and reconstruction error (MSE) between \code{rLifting} and other packages
(wavethresh, wavelets) using Haar wavelet.
}
\keyword{datasets}
