% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.apply_filter_lifting}
\alias{.apply_filter_lifting}
\title{Aplica filtro lifting com suporte a bordas}
\usage{
.apply_filter_lifting(x, coeffs, start_idx = 0, extension = "symmetric")
}
\arguments{
\item{x}{Sinal input.}

\item{coeffs}{Coeficientes do filtro P ou U.}

\item{start_idx}{Offset relativo.}

\item{extension}{Modo de extensao ("symmetric", "periodic", "zero").}
}
\description{
Aplica filtro lifting com suporte a bordas
}
\keyword{internal}
