% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifting_scheme.R
\name{lifting_scheme}
\alias{lifting_scheme}
\title{Construtor para o Esquema de Lifting (Lifting Scheme)}
\usage{
lifting_scheme(wavelet = "haar", custom_steps = NULL, custom_norm = NULL)
}
\arguments{
\item{wavelet}{Nome da ondaleta (string). Ex: "haar", "db2", "cdf97".}

\item{custom_steps}{Lista de passos (opcional).}

\item{custom_norm}{Vetor de normalizacao (opcional).}
}
\value{
Um objeto da classe \code{lifting_scheme}.
}
\description{
Cria um objeto S3 contendo os passos de predicao e atualizacao.
}
