% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptative_threshold.R
\name{plot.adaptive_thresholds}
\alias{plot.adaptive_thresholds}
\title{Plot method for Adaptive Thresholds}
\usage{
\method{plot}{adaptive_thresholds}(x, ...)
}
\arguments{
\item{x}{Object of class \code{adaptive_thresholds}.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Plot method for Adaptive Thresholds
}
