% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifting_scheme.R
\name{plot.lifting_scheme}
\alias{plot.lifting_scheme}
\title{Plot method for Lifting Scheme}
\usage{
\method{plot}{lifting_scheme}(x, ...)
}
\arguments{
\item{x}{An object of class \code{lifting_scheme}.}

\item{...}{Additional arguments passed to \code{visualize_wavelet_basis}.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Plot method for Lifting Scheme
}
