% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwt.R
\name{plot.lwt}
\alias{plot.lwt}
\title{Plot method for LWT Decomposition}
\usage{
\method{plot}{lwt}(x, ...)
}
\arguments{
\item{x}{An object of class \code{lwt}.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Plot method for LWT Decomposition
}
