% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptative_threshold.R
\name{print.adaptive_thresholds}
\alias{print.adaptive_thresholds}
\title{Print method for Adaptive Thresholds}
\usage{
\method{print}{adaptive_thresholds}(x, ...)
}
\arguments{
\item{x}{Object of class \code{adaptive_thresholds}.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for Adaptive Thresholds
}
