% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifting_scheme.R
\name{print.lifting_scheme}
\alias{print.lifting_scheme}
\title{Print method}
\usage{
\method{print}{lifting_scheme}(x, ...)
}
\arguments{
\item{x}{object of class \code{lifting_scheme}.}

\item{...}{additional arguments.}
}
\value{
Invisibly returns \code{NULL}. Called for side effects (printing).
}
\description{
Print method
}
