% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwt.R
\name{print.lwt}
\alias{print.lwt}
\title{Print method for LWT}
\usage{
\method{print}{lwt}(x, ...)
}
\arguments{
\item{x}{An object of class lwt.}

\item{...}{Additional arguments.}
}
\description{
Print method for LWT
}
