% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{print.wavelet_diagnosis}
\alias{print.wavelet_diagnosis}
\title{Print method for Wavelet Diagnosis}
\usage{
\method{print}{wavelet_diagnosis}(x, ...)
}
\arguments{
\item{x}{Object of class \code{wavelet_diagnosis}.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for Wavelet Diagnosis
}
