% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime_denoising.R
\name{print.wavelet_stream}
\alias{print.wavelet_stream}
\title{Print method for Wavelet Stream Processor}
\usage{
\method{print}{wavelet_stream}(x, ...)
}
\arguments{
\item{x}{Object of class \code{wavelet_stream}.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for Wavelet Stream Processor
}
