% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{rLifting}
\alias{rLifting-package}
\alias{rLifting}
\title{rLifting: Real-Time Signal Denoising via Lifting Scheme}
\description{
A package for wavelet transforms using the Lifting Scheme, focused on
real-time (causal) applications, adaptive thresholding, and custom
wavelet construction.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
