% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding.R
\name{threshold_hard}
\alias{threshold_hard}
\title{Hard Thresholding}
\usage{
threshold_hard(x, lambda)
}
\arguments{
\item{x}{Vetor de coeficientes (detalhes).}

\item{lambda}{Limiar (threshold) positivo.}
}
\value{
Vetor processado.
}
\description{
Zera coeficientes abaixo do limiar, mantem os demais inalterados.
Equivalente a "keep or kill".
}
