% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{visualize_wavelet_basis}
\alias{visualize_wavelet_basis}
\title{Visualize Basis Functions (Scaling and Wavelet)}
\usage{
visualize_wavelet_basis(scheme, plot = TRUE, levels = 8)
}
\arguments{
\item{scheme}{Object of class \code{lifting_scheme}.}

\item{plot}{Boolean.}

\item{levels}{Number of cascade levels.}
}
\description{
Plots the waveform by iterating the reconstruction over several levels.
}
