% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnose_wavelet}
\alias{diagnose_wavelet}
\title{Diagnostico Completo de Ondaleta}
\usage{
diagnose_wavelet(wavelet_name, config, verbose = TRUE)
}
\arguments{
\item{wavelet_name}{Nome ou objeto.}

\item{config}{Lista de configuracao (is_ortho, vm_degrees, max_taps).}

\item{verbose}{Imprimir no console?}
}
\description{
Executa bateria de testes fisicos.
}
