% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnose_wavelet}
\alias{diagnose_wavelet}
\title{Complete Wavelet Diagnosis}
\usage{
diagnose_wavelet(wavelet_name, config, verbose = TRUE)
}
\arguments{
\item{wavelet_name}{Name string or a \code{lifting_scheme} object.}

\item{config}{Configuration list (is_ortho, vm_degrees, max_taps).}

\item{verbose}{Print results to console?}
}
\value{
(Invisible) A list containing the results of each test.
}
\description{
Runs a battery of physical and mathematical tests on a wavelet.
}
