% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.apply_filter_lifting}
\alias{.apply_filter_lifting}
\title{Applies lifting filter with boundary support (C++ Wrapper)}
\usage{
.apply_filter_lifting(x, coeffs, start_idx = 0, extension = "symmetric")
}
\arguments{
\item{x}{Input signal.}

\item{coeffs}{Coefficients of filter P or U.}

\item{start_idx}{Relative offset.}

\item{extension}{Extension mode.}
}
\description{
This function bridges the R logic to the compiled C++ core for speed.
}
\keyword{internal}
