% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.pad_signal}
\alias{.pad_signal}
\title{Tratamento de Padding/Extensao de Sinais}
\usage{
.pad_signal(x, n_left, n_right, mode = "symmetric")
}
\arguments{
\item{x}{Sinal de entrada.}

\item{n_left}{Amostras a esquerda.}

\item{n_right}{Amostras a direita.}

\item{mode}{Modo de extensao: "symmetric", "periodic", "zero".}
}
\description{
Tratamento de Padding/Extensao de Sinais
}
\keyword{internal}
