% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilwt.R
\name{ilwt}
\alias{ilwt}
\title{Inverse Lifting Wavelet Transform}
\usage{
ilwt(lwt_obj, scheme = NULL)
}
\arguments{
\item{lwt_obj}{Objeto da classe 'lwt' gerado pela funcao lwt().}

\item{scheme}{(Opcional) Objeto lifting_scheme. Se NULL, usa o do lwt_obj.}
}
\value{
Vetor numerico com o sinal reconstruido.
}
\description{
Reconstroi o sinal original a partir dos coeficientes de ondaleta.
}
\examples{
s = c(1, 2, 3, 4)
sch = lifting_scheme("haar")
fwd = lwt(s, sch)
rec = ilwt(fwd)
print(rec) # Deve ser igual a s
}
