% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifting_scheme.R
\name{lifting_scheme}
\alias{lifting_scheme}
\title{Lifting Scheme Constructor}
\usage{
lifting_scheme(wavelet = "haar", custom_steps = NULL, custom_norm = NULL)
}
\arguments{
\item{wavelet}{Wavelet name (string).
Options: "haar", "db2", "cdf53", "cdf97", "dd4", "lazy".}

\item{custom_steps}{List of custom steps (optional).
If provided, ignores internal lookup.}

\item{custom_norm}{Normalization vector (optional).}
}
\value{
An object of class \code{lifting_scheme}.
}
\description{
Creates an S3 object containing the prediction (P) and update (U) steps
required for the Lifting Transform.
}
