% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwt.R
\name{lwt}
\alias{lwt}
\title{Lifting Wavelet Transform (Forward)}
\usage{
lwt(signal, scheme, levels = 1, extension = "symmetric")
}
\arguments{
\item{signal}{Vetor numerico contendo o sinal.}

\item{scheme}{Objeto da classe 'lifting_scheme'.}

\item{levels}{Inteiro. Numero de niveis de decomposicao.}

\item{extension}{Modo de tratamento de borda: "symmetric" (padrao),
"periodic", "zero".}
}
\value{
Objeto S3 da classe 'lwt'. Contem:
\item{coeffs}{Lista com detalhes (d1..dn) e aproximacao (an).}
\item{scheme}{O esquema utilizado.}
}
\description{
Executa a transformada wavelet via esquema de lifting.
Suporta decomposicao multinivel.
}
\examples{
data = c(1, 2, 3, 4, 5, 6, 7, 8)
sch = lifting_scheme("haar")
res = lwt(data, sch, levels = 2)
print(res)
}
