% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime_denoising.R
\name{new_wavelet_stream}
\alias{new_wavelet_stream}
\title{Cria um Processador de Fluxo (Stream) Wavelet Adaptativo}
\usage{
new_wavelet_stream(
  scheme,
  window_size = 256,
  alpha = 0.3,
  beta = 1.2,
  method = "semisoft",
  extension = "symmetric"
)
}
\arguments{
\item{scheme}{Objeto \code{lifting_scheme}.}

\item{window_size}{Tamanho da janela (W). Ex: 256.}

\item{alpha}{Parametro de decaimento do threshold (Eq 9).}

\item{beta}{Fator de ganho do threshold (Eq 9).}

\item{method}{Metodo de encolhimento: "hard", "soft", "semisoft".}

\item{extension}{Tratamento de borda ('symmetric', 'periodic').}
}
\value{
Funcao processadora stream(x).
}
\description{
Gera uma funcao stateful que implementa o metodo de Liu et al. completo:
Janela Movel + Decomposicao Lifting + Threshold Recursivo Adaptativo.
}
