% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifting_scheme.R
\name{print.lifting_scheme}
\alias{print.lifting_scheme}
\title{Print method}
\usage{
\method{print}{lifting_scheme}(x, ...)
}
\arguments{
\item{x}{lifting_scheme object.}

\item{...}{additional arguments.}
}
\description{
Print method
}
