% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwt.R
\name{print.lwt}
\alias{print.lwt}
\title{Print method para LWT}
\usage{
\method{print}{lwt}(x, ...)
}
\arguments{
\item{x}{Objeto da classe lwt.}

\item{...}{Argumentos adicionais (nao utilizados).}
}
\description{
Print method para LWT
}
