% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{rLifting}
\alias{rLifting-package}
\alias{rLifting}
\title{rLifting: High-Performance Wavelet Lifting Transforms}
\description{
A unified framework for Wavelet Transforms using the Lifting Scheme.
It provides robust tools for offline signal analysis and functional data
analysis (FDA), while also enabling high-performance causal processing
for real-time applications via a specialized C++ core.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mkyou/rLifting}
  \item Report bugs at \url{https://github.com/mkyou/rLifting/issues}
}

}
\author{
\strong{Maintainer}: Moises da Silva \email{moisesdff8@gmail.com}

}
