% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding.R
\name{threshold}
\alias{threshold}
\title{General Thresholding Wrapper}
\usage{
threshold(x, lambda, method = "soft")
}
\arguments{
\item{x}{Input vector.}

\item{lambda}{Threshold value.}

\item{method}{Method: "hard", "soft" or "semisoft".}
}
\value{
Numeric vector of the same length as \code{x} with thresholded coefficients.
}
\description{
General Thresholding Wrapper
}
