% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding.R
\name{threshold}
\alias{threshold}
\title{General Thresholding Wrapper}
\usage{
threshold(x, lambda, method = "soft")
}
\arguments{
\item{x}{Input vector.}

\item{lambda}{Threshold.}

\item{method}{"hard", "soft" or "semisoft".}
}
\description{
General Thresholding Wrapper
}
