% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding.R
\name{threshold}
\alias{threshold}
\title{Wrapper Geral de Thresholding}
\usage{
threshold(x, lambda, method = "soft")
}
\arguments{
\item{x}{Vetor de entrada.}

\item{lambda}{Limiar.}

\item{method}{"hard", "soft" ou "semisoft".}
}
\description{
Wrapper Geral de Thresholding
}
