% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding.R
\name{threshold_hard}
\alias{threshold_hard}
\title{Hard Thresholding}
\usage{
threshold_hard(x, lambda)
}
\arguments{
\item{x}{Vector of coefficients (details).}

\item{lambda}{Positive threshold value.}
}
\value{
Processed vector.
}
\description{
Sets coefficients below the threshold to zero, keeping others unchanged.
Known as the "keep or kill" policy.
}
