% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding.R
\name{threshold_semisoft}
\alias{threshold_semisoft}
\title{Semisoft Shrinkage (Hyperbolic)}
\usage{
threshold_semisoft(x, lambda)
}
\arguments{
\item{x}{Vector of coefficients.}

\item{lambda}{Positive threshold value.}
}
\value{
Processed vector.
}
\description{
Implementation based on Liu et al. (2014).
Combines the stability of Soft Thresholding with the amplitude
precision of Hard Thresholding.
Function: \code{sign(x) * sqrt(x^2 - lambda^2)} for values above lambda.
}
