% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding.R
\name{threshold_semisoft}
\alias{threshold_semisoft}
\title{Semisoft Shrinkage (Hyperbolic)}
\usage{
threshold_semisoft(x, lambda)
}
\arguments{
\item{x}{Vetor de coeficientes.}

\item{lambda}{Limiar positivo.}
}
\value{
Vetor processado.
}
\description{
Implementacao baseada em Liu et al. (2014).
Combina a estabilidade do Soft com a precisao de amplitude do Hard.
Funcao: sign(x) * sqrt(x^2 - lambda^2) para |x| > lambda.
}
