% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding.R
\name{threshold_soft}
\alias{threshold_soft}
\title{Soft Thresholding}
\usage{
threshold_soft(x, lambda)
}
\arguments{
\item{x}{Vector of coefficients.}

\item{lambda}{Positive threshold value.}
}
\value{
Processed vector.
}
\description{
Sets coefficients below the threshold to zero and shrinks others
towards zero.
Reduces noise variance but introduces amplitude bias.
}
